Implement flask jwt extended

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 
 from models.item import ItemModel
 
@@ -34,7 +34,12 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -57,11 +62,15 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_required(optional=True)
     def get(self):
-        # return {
-        #     'items': [item.json() for item in ItemModel.query.all()]
-        # }
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.query.all()]
+
+        if user_id:
+            return {'items': items}
 
         return {
-            'items': list(map(lambda item: item.json(), ItemModel.find_all()))
+            'items': [item['name'] for item in items],
+            'message': 'More data available if you log in'
         }
